requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
regex==2024.9.11